requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
cachetools==4.2.4; python_version ~= '3.5'
certifi==2021.10.8
charset-normalizer==2.0.7; python_version >= '3'
click==8.0.3; python_version >= '3.6'
flask==2.0.2
google-api-core[grpc]==2.2.2; python_version >= '3.6'
google-auth==2.3.3; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
google-cloud-bigquery==2.30.1
google-cloud-core==2.2.1; python_version >= '3.6'
google-crc32c==1.3.0; python_version >= '3.6'
google-resumable-media==2.1.0; python_version >= '3.6'
googleapis-common-protos==1.53.0; python_version >= '3.6'
grpcio-status==1.41.1
grpcio==1.41.1
idna==3.3; python_version >= '3'
itsdangerous==2.0.1; python_version >= '3.6'
jinja2==3.0.3; python_version >= '3.6'
markupsafe==2.0.1; python_version >= '3.6'
packaging==21.2; python_version >= '3.6'
proto-plus==1.19.8; python_version >= '3.6'
protobuf==3.19.1; python_version >= '3.5'
pyasn1-modules==0.2.8
pyasn1==0.4.8
pyparsing==2.4.7; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'
python-dateutil==2.8.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
requests==2.26.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
rsa==4.7.2; python_version >= '3.6'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
urllib3==1.26.7; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4' and python_version < '4'
werkzeug==2.0.2; python_version >= '3.6'
